Se actualizo los JPA

+++ src/main/java/pe/edu/utp/i2/citasmedicas/dao/api/PacienteDaoAPI.java
@@ -1,9 +1,8 @@
 package pe.edu.utp.i2.citasmedicas.dao.api;
 
 import org.springframework.data.repository.CrudRepository;
-
 import pe.edu.utp.i2.citasmedicas.model.Paciente;
 
-public interface PacienteDaoAPI extends CrudRepository<Paciente, Long> {
+public interface PacienteDaoAPI extends CrudRepository<Paciente, Integer> {
 
 }
